feat: Refactor API structure and move implementation to forge_api module
