+++ crates/forge_api/src/forge_api.rs
@@ -1,0 +1,104 @@
+use std::path::Path;
+use std::sync::Arc;
+
+use anyhow::Result;
+use forge_domain::*;
+use forge_infra::ForgeInfra;
+use forge_services::{EnvironmentService, ForgeApp, Infrastructure};
+use forge_stream::MpscStream;
+use serde_json::Value;
+
+use crate::executor::ForgeExecutorService;
+use crate::loader::ForgeLoaderService;
+use crate::suggestion::ForgeSuggestionService;
+use crate::API;
+
+pub struct ForgeAPI<F> {
+    app: Arc<F>,
+    executor_service: ForgeExecutorService<F>,
+    suggestion_service: ForgeSuggestionService<F>,
+    loader: ForgeLoaderService<F>,
+}
+
+impl<F: App + Infrastructure> ForgeAPI<F> {
+    pub fn new(app: Arc<F>) -> Self {
+        Self {
+            app: app.clone(),
+            executor_service: ForgeExecutorService::new(app.clone()),
+            suggestion_service: ForgeSuggestionService::new(app.clone()),
+            loader: ForgeLoaderService::new(app.clone()),
+        }
+    }
+}
+
+impl ForgeAPI<ForgeApp<ForgeInfra>> {
+    pub fn init(restricted: bool) -> Self {
+        let infra = Arc::new(ForgeInfra::new(restricted));
+        let app = Arc::new(ForgeApp::new(infra));
+        ForgeAPI::new(app)
+    }
+}
+
+#[async_trait::async_trait]
+impl<F: App + Infrastructure> API for ForgeAPI<F> {
+    async fn suggestions(&self) -> Result<Vec<File>> {
+        self.suggestion_service.suggestions().await
+    }
+
+    async fn tools(&self) -> Vec<ToolDefinition> {
+        self.app.tool_service().list()
+    }
+
+    async fn models(&self) -> Result<Vec<Model>> {
+        Ok(self.app.provider_service().models().await?)
+    }
+
+    async fn chat(
+        &self,
+        chat: ChatRequest,
+    ) -> anyhow::Result<MpscStream<Result<AgentMessage<ChatResponse>, anyhow::Error>>> {
+        Ok(self.executor_service.chat(chat).await?)
+    }
+
+    async fn init(&self, workflow: Workflow) -> anyhow::Result<ConversationId> {
+        self.app.conversation_service().create(workflow).await
+    }
+
+    fn environment(&self) -> Environment {
+        self.app.environment_service().get_environment().clone()
+    }
+
+    async fn load(&self, path: Option<&Path>) -> anyhow::Result<Workflow> {
+        self.loader.load(path).await
+    }
+
+    async fn conversation(
+        &self,
+        conversation_id: &ConversationId,
+    ) -> anyhow::Result<Option<Conversation>> {
+        self.app.conversation_service().find(conversation_id).await
+    }
+
+    async fn get_variable(
+        &self,
+        conversation_id: &ConversationId,
+        key: &str,
+    ) -> anyhow::Result<Option<Value>> {
+        self.app
+            .conversation_service()
+            .get_variable(conversation_id, key)
+            .await
+    }
+
+    async fn set_variable(
+        &self,
+        conversation_id: &ConversationId,
+        key: String,
+        value: Value,
+    ) -> anyhow::Result<()> {
+        self.app
+            .conversation_service()
+            .set_variable(conversation_id, key, value)
+            .await
+    }
+}
